+++ lightly_studio/src/lightly_studio/core/dataset_query/__init__.py
@@ -1,7 +1,7 @@
 from .boolean_expression import AND, NOT, OR
 from .dataset_query import DatasetQuery
+from .image_sample_field import ImageSampleField
 from .order_by import OrderByExpression, OrderByField
-from .sample_field import SampleField
 from .video_sample_field import VideoSampleField
 
 __all__ = [
@@ -9,8 +9,8 @@ __all__ = [
     "NOT",
     "OR",
     "DatasetQuery",
+    "ImageSampleField",
     "OrderByExpression",
     "OrderByField",
-    "SampleField",
     "VideoSampleField",
 ]