+++ lightly_studio/src/lightly_studio/core/dataset_query/image_sample_field.py
@@ -13,18 +13,18 @@ from lightly_studio.core.dataset_query.tags_expression import TagsAccessor
 from lightly_studio.models.image import ImageTable
 
 
-class SampleField:
+class ImageSampleField:
     """Providing access to predefined sample fields for queries.
 
     It is used for the `query.match(...)` and `query.order_by(...)` methods of the
     `DatasetQuery` class.
 
     ```python
-    from lightly_studio.core.dataset_query import SampleField, OrderByField
+    from lightly_studio.core.dataset_query import ImageSampleField, OrderByField
 
     query = dataset.query()
-    query.match(SampleField.tags.contains("cat"))
-    query.order_by(OrderByField(SampleField.file_path_abs))
+    query.match(ImageSampleField.tags.contains("cat"))
+    query.order_by(OrderByField(ImageSampleField.file_path_abs))
     samples = query.to_list()
     ```
     """