+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Reduced the package size by using opencv-python-headless.
 - AnnotationLabelTable is now linked to a dataset.
 - `lightly_studio.Dataset` class has been renamed to `lightly_studio.ImageDataset`.
+- Renamed `SampleField` to `ImageSampleField`.
 
 ### Deprecated